separated S3 tests from the others
